fix ci failure
