+++ e2e/support/helpers/component-testing-sdk/component-embedding-sdk-helpers.tsx
@@ -2,11 +2,9 @@ import {
   MetabaseProvider,
   type MetabaseProviderProps,
 } from "@metabase/embedding-sdk-react";
-import * as jose from "jose";
 import type { JSX } from "react";
 import React from "react";
 
-import { USERS } from "e2e/support/cypress_data";
 import { signInAsAdminAndEnableEmbeddingSdkForE2e } from "e2e/support/helpers/e2e-embedding-sdk-helpers";
 import { ThemeProvider } from "metabase/ui";
 
@@ -19,45 +17,6 @@ export const JWT_SHARED_SECRET =
 
 export const DEFAULT_SDK_AUTH_PROVIDER_CONFIG = {
   metabaseInstanceUrl: METABASE_INSTANCE_URL,
-};
-
-export const mockAuthProviderAndJwtSignIn = (user = USERS.admin) => {
-  cy.intercept("GET", `${AUTH_PROVIDER_URL}**`, async (req) => {
-    try {
-      const url = new URL(req.url);
-      const responseParam = url.searchParams.get("response");
-
-      // Return error if response is not json
-      if (responseParam !== "json") {
-        req.reply({
-          statusCode: 400,
-          body: { error: "Invalid response parameter. Expected response=json" },
-        });
-        return;
-      }
-
-      const secret = new TextEncoder().encode(JWT_SHARED_SECRET);
-      const jwt = await new jose.SignJWT({
-        email: user.email,
-        first_name: user.first_name,
-        last_name: user.last_name,
-        exp: Math.round(Date.now() / 1000) + 60 * 10, // 10 minutes expiration
-      })
-        .setProtectedHeader({ alg: "HS256" })
-        .sign(secret);
-
-      req.reply({
-        statusCode: 200,
-        body: { jwt },
-      });
-    } catch (error: any) {
-      console.warn("SDK auth error:", error);
-      req.reply({
-        statusCode: 500,
-        body: error.message,
-      });
-    }
-  }).as("jwtProvider");
 };
 
 export interface MountSdkContentOptions {
